feat(xgon): change pending tx metric logic to singe file
